+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+owncloud-news (4.1.1)
+* **Enhancement**: shortcut **a** scrolls to the currently active feed in the navigation
+* **Enhancement**: next/previous feed/folder shortcuts now scroll to the active entry
+
 owncloud-news (4.1.0)
 * **Backwards incompatible change**: Time used for updating when using the Python updater is now subtracted from the given interval meaning: if you specify 30 seconds as interval and the update takes 25 seconds, it will sleep for 5 seconds before running the next update
 * **Enhancement**: Show a message if ajax cron is used to inform people of improperly configured cron